+++ server/copy.sh
@@ -1,0 +1,3 @@
+cp -Rvf src/ ../client/server
+cp -vf package.json ../client/server
+cp -Rvf node_modules/ ../client/server