+++ client/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "harbour" extension will be documented in this file.
 ## Planned
 - Show matches on 'if-else-endif', 'for-exit-loop-next' etc...
 
+## 0.4.1
+ - send symbol kind in the correct way to have icons
+
+## 0.4.0
+ - added Language server
+ - Added workspace symbol provider
+
 ## 0.3.5
 - fixed crashes in debugger (need recompile the library too)